chore: Apply typescript
