+++ src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css';
-import {Routes, Route} from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 import Home from './routes/Home';
 import About from './routes/About';
 import SignIn from './routes/SignIn';
