--- src/Components/Copyright.js
@@ -1,18 +1,0 @@
-import Typography from '@mui/material/Typography';
-import Link from '@mui/material/Link';
-import * as React from 'react';
-
-function Copyright(props) {
-  return (
-      <Typography variant="body2" color="text.secondary" align="center" {...props}>
-        {'Copyright © '}
-        <Link color="inherit" href="https://mui.com/">
-          Your Website
-        </Link>{' '}
-        {new Date().getFullYear()}
-        {'.'}
-      </Typography>
-  );
-}
-
-export default Copyright;
