--- src/routes/About.js
@@ -1,9 +1,0 @@
-function About() {
-  return (
-      <div>
-        <h2>Home</h2>
-      </div>
-  );
-}
-
-export default About;
