+++ src/routes/Home.tsx
@@ -1,0 +1,9 @@
+function Home() {
+  return (
+    <div>
+      <h2>Home</h2>
+    </div>
+  );
+}
+
+export default Home;
